% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FimsUtils.R
\name{authenticate}
\alias{authenticate}
\title{Must call this first!
authenticate, necessary to fetch data which is private}
\usage{
authenticate(user, pass)
}
\description{
Must call this first!
authenticate, necessary to fetch data which is private
}

