% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FimsUtils.R
\name{listEntities}
\alias{listEntities}
\title{Get a list of entities (i.e. tables) available to query}
\usage{
listEntities(projectId = NA)
}
\arguments{
\item{projectId}{(optional) The project to fetch entities for. If not provided, the network entities will be returned.}
}
\description{
Get a list of entities (i.e. tables) available to query
}
\examples{
\dontrun{
entities <- listEntities(projectId)
entities <- listEntities()
}
}
