% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FimsUtils.R
\name{listLoci}
\alias{listLoci}
\title{Get a list of loci that are stored in FASTA format directly in GEOME (not in the SRA)}
\usage{
listLoci()
}
\description{
Get a list of loci that are stored in FASTA format directly in GEOME (not in the SRA)
}
\examples{
\dontrun{
markers <- listLoci()
}
}
