% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FimsUtils.R
\name{queryFasta}
\alias{queryFasta}
\title{fetch Fasta sequences from the dipnet database}
\usage{
queryFasta(marker, expeditions = list(), filters = list())
}
\arguments{
\item{marker}{the marker to fetch}

\item{expeditions}{list of expeditions to include in the query. The default is all expeditions}

\item{filters}{list of "column name":"value" pairs to include in the query. The special column name "_all" will do
a full text search in all columns for the value}
}
\description{
fetch Fasta sequences from the dipnet database
}

