% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FimsUtils.R
\name{listExpeditions}
\alias{listExpeditions}
\title{get a list of expeditions for a project}
\usage{
listExpeditions(projectId)
}
\description{
get a list of expeditions for a project
}
\examples{
\dontrun{
expeditions <- listExpeditions(projectId)
}
}
