% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FimsUtils.R
\name{queryFasta}
\alias{queryFasta}
\title{fetch Fasta sequences from the geome-db database}
\usage{
queryFasta(marker, expeditions = list(), query = "")
}
\arguments{
\item{marker}{the marker to fetch}

\item{expeditions}{list of expeditions to include in the query. The default is all expeditions}

\item{query}{FIMS Query DSL \url{http://fims.readthedocs.io/en/latest/fims/query.html} query string.
Ex. '+locality:fuzzy +country:"exact phrase"'}
}
\description{
fetch Fasta sequences from the geome-db database
}
\examples{
\dontrun{
fasta <- queryFasta("C01", expeditions=list("TEST"), query="+yearCollected:2008")
}
}

