% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FimsUtils.R
\name{queryMetadata}
\alias{queryMetadata}
\title{fetch the FimsMetadata from the dipnet database}
\usage{
queryMetadata(expeditions = list(), filters = list(), names = NULL)
}
\arguments{
\item{expeditions}{list of expeditions to include in the query. The default is all expeditions}

\item{filters}{list of "column name":"value" pairs to include in the query. The special column name "_all" will do
a full text search in all columns for the value}

\item{names}{list of column names to include in the data.frame results}
}
\description{
fetch the FimsMetadata from the dipnet database
}

